fix(app): move not-found to root so unmatched routes render 404 chrome
After splitting chrome into the (public) route group, the group's
not-found.tsx stopped handling globally unmatched URLs — Next.js only
uses app/not-found.tsx for those. Move the 404 page up to the root and
include Nav/Footer inline so unmatched routes still render the full UI.

Fixes e2e 404 tests that assert a "Back to Home" link inside <main>.
